requirements.txt
streamlit
crewai
python-dotenv
duckduckgo-search
pandas
plotly
graphviz
